py/pkg/status-grpc/pyproject.toml
[tool.poetry]
name = "status-grpc"
version = "0.1.1"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.43.0"
google = "^3.0.0"
protobuf = "^3.19.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
